CRUD refactored and Pydantic Model Bug fix

+++ backend/app.py
@@ -1,12 +1,12 @@
-from fastapi import FastAPI, Request, Depends, status
-from fastapi.middleware.cors import CORSMiddleware
-from fastapi.responses import JSONResponse
+from typing import List
 
+from fastapi import FastAPI, Depends, status, HTTPException
+from fastapi.middleware.cors import CORSMiddleware
 from sqlalchemy.orm import Session
 
 import models
 from models import Todo
-from pydantic_models.Todo import TodoCreate
+from pydantic_models.Todo import TodoCreate, TodoResponse
 
 from database import SessionLocal, engine
 
@@ -32,42 +32,43 @@ def get_db():
     db.close()
 
 # CRUD for API
-@app.get("/api/todo/list")
+# List Todos
+@app.get("/api/todo/list", response_model=List[TodoResponse], status_code=status.HTTP_200_OK)
 def get_todos(db: Session = Depends(get_db)):
-  todos = db.query(Todo).all()
-  todos_data = [todo.__dict__ for todo in todos]
-  # Exclude metadata for serialization
-  todos_data = [{k: v for k, v in todo.items() if k != "_sa_instance_state"} for todo in todos_data]
-  
-  return JSONResponse(status_code=status.HTTP_200_OK, content=todos_data)
+    return db.query(Todo).all()
 
-@app.post("/api/todo/new")
+# New Todo
+@app.post("/api/todo/new", response_model=TodoResponse, status_code=status.HTTP_201_CREATED)
 def new_todo(todo_in: TodoCreate, db: Session = Depends(get_db)):
     print(">>>> Attempting to add todo: ", todo_in)
-    todo = Todo(title=todo_in.title)
-    db.add(todo)
+    new_todo = Todo(title=todo_in.title)
+    db.add(new_todo)
     db.commit()
-    db.refresh(todo)
-    
+    db.refresh(new_todo)
+
+    return new_todo
+
+# Update Todo
+@app.patch("/api/update/{todo_id}", response_model=TodoResponse)
+def update_todo(todo_id: int, db: Session = Depends(get_db)):
+    print(">>>> Attempting to update todo: ", todo_id)
+    todo = db.query(Todo).filter(Todo.id == todo_id).first()
+    if not todo:
+        raise HTTPException(status_code=404, detail="Todo not found")
+    todo.complete = not todo.complete
+    db.commit()
+
     return todo
 
-@app.patch("/api/update/{todo_id}")
-def update_todo(request: Request, todo_id: int, db: Session = Depends(get_db)):
-  print(">>>> Attempting to update todo id: ", todo_id)
-  todo = db.query(Todo).filter(Todo.id == todo_id).first()
 
-  todo.complete = not todo.complete
-  todo_data = todo.__dict__
-  todo_data = {k: v for k, v in todo_data.items() if k != "_sa_instance_state"}
-  db.commit()
-  
-  return JSONResponse(status_code=status.HTTP_200_OK, content=todo_data)
+# Delete todo
+@app.delete("/api/delete/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
+def delete_todo(todo_id: int, db: Session = Depends(get_db)):
+    print(">>>> Attempting to delete todo: ", todo_id)
+    todo = db.query(Todo).filter(Todo.id == todo_id).first()
+    if not todo:
+        raise HTTPException(status_code=404, detail="Todo not found")
+    db.delete(todo)
+    db.commit()
 
-@app.delete("/api/delete/{todo_id}")
-def delete_todo(request: Request, todo_id: int, db: Session = Depends(get_db)):
-  print(">>>> Attempting to delete todo id: ", todo_id)
-  todo = db.query(Todo).filter(Todo.id == todo_id).first()
-  db.delete(todo)
-  db.commit()
-
-  return JSONResponse(status_code=status.HTTP_204_NO_CONTENT, content={"message": "Todo deleted"})
+    return {"message": "Todo deleted"}
